// src/types/user.ts
export type User = {
  id?: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  jobArea: string;
  jobTitle: string;
  jobType: string;
  jobDescription: string;
  avatar: string;
  country: string;
  city: string;
  streetAdress: string;
  zipCode: string;
  state: string;
};